# Faster prefix sum for bitsPerValue up to 9.