--- a/lucene/core/src/generated/checksums/generateForUtil.json
+++ b/lucene/core/src/generated/checksums/generateForUtil.json
@@ -1,4 +1,4 @@
 {
-    "lucene/core/src/java/org/apache/lucene/codecs/lucene99/ForUtil.java": "1292ad354d255b1272ffd3db684aa2ddb2bc49ec",
-    "lucene/core/src/java/org/apache/lucene/codecs/lucene99/gen_ForUtil.py": "ab7b63a1b73986cc04e43de1c8f474b97aef5116"
+    "lucene/core/src/java/org/apache/lucene/codecs/lucene99/ForUtil.java": "679bac572b76b8e6a0cc14ab9c74a56f497a2203",
+    "lucene/core/src/java/org/apache/lucene/codecs/lucene99/gen_ForUtil.py": "ac1bc462ee4c889dbd24f741abaf64498ae132d6"
 }