--- a/lucene/core/src/generated/checksums/generateForUtil.json
+++ b/lucene/core/src/generated/checksums/generateForUtil.json
@@ -1,4 +1,4 @@
 {
-    "lucene/core/src/java/org/apache/lucene/codecs/lucene99/ForUtil.java": "679bac572b76b8e6a0cc14ab9c74a56f497a2203",
-    "lucene/core/src/java/org/apache/lucene/codecs/lucene99/gen_ForUtil.py": "ac1bc462ee4c889dbd24f741abaf64498ae132d6"
+    "lucene/core/src/java/org/apache/lucene/codecs/lucene99/ForUtil.java": "e0c3acef3bca93023611d43502a1e9c40f490701",
+    "lucene/core/src/java/org/apache/lucene/codecs/lucene99/gen_ForUtil.py": "af6573d06dffd0ae2ecee02708c46a72526fbc8b"
 }