--- a/lucene/core/src/generated/checksums/generateForUtil.json
+++ b/lucene/core/src/generated/checksums/generateForUtil.json
@@ -1,4 +1,4 @@
 {
-    "lucene/core/src/java/org/apache/lucene/codecs/lucene99/ForUtil.java": "e0c3acef3bca93023611d43502a1e9c40f490701",
-    "lucene/core/src/java/org/apache/lucene/codecs/lucene99/gen_ForUtil.py": "af6573d06dffd0ae2ecee02708c46a72526fbc8b"
+    "lucene/core/src/java/org/apache/lucene/codecs/lucene99/ForUtil.java": "ad68a0d2be5eef4117086893ea9656e84376dd85",
+    "lucene/core/src/java/org/apache/lucene/codecs/lucene99/gen_ForUtil.py": "d8a9d5fca1d4d02769cc8475bd08cccd1c3f7876"
 }