In order to speed up computing prefix sums, we currently pack two deltas in a long and then sum up longs. This allows us to sum up two deltas in a single instruction. We can do better for numbers of bits per value up to 9: since `128 * 2^9 = 2^16`, we can actually compute the prefix sum on only 16 bits per value, summing up 4 deltas at once.