@jpountz 2023-11-24
luceneutil doesn't see a noticeable difference (all p-values are high) but the micro-benchmark that is attached to this PR seems to see an improvement:

```
main

Benchmark                            (bpv)   Mode  Cnt   Score   Error   Units
ForUtilBenchmark.decodeAndPrefixSum      6  thrpt   25  18.762 ± 0.739  ops/us
ForUtilBenchmark.decodeAndPrefixSum      7  thrpt   25  18.075 ± 0.220  ops/us
ForUtilBenchmark.decodeAndPrefixSum      8  thrpt   25  21.040 ± 0.285  ops/us
ForUtilBenchmark.decodeAndPrefixSum      9  thrpt   25  16.790 ± 0.896  ops/us
ForUtilBenchmark.decodeAndPrefixSum     10  thrpt   25  17.441 ± 1.260  ops/us
ForUtilBenchmark.decodeAndPrefixSum     11  thrpt   25  16.697 ± 0.883  ops/us

PR

Benchmark                            (bpv)   Mode  Cnt   Score   Error   Units
ForUtilBenchmark.decodeAndPrefixSum      6  thrpt   25  19.171 ± 0.277  ops/us
ForUtilBenchmark.decodeAndPrefixSum      7  thrpt   25  18.875 ± 0.203  ops/us
ForUtilBenchmark.decodeAndPrefixSum      8  thrpt   25  22.075 ± 0.497  ops/us
ForUtilBenchmark.decodeAndPrefixSum      9  thrpt   25  18.689 ± 0.792  ops/us
ForUtilBenchmark.decodeAndPrefixSum     10  thrpt   25  17.696 ± 0.252  ops/us
ForUtilBenchmark.decodeAndPrefixSum     11  thrpt   25  16.623 ± 0.856  ops/us
```