@jpountz 2023-11-24
Actually we can do even better by better tuning the disk layout for the prefix sum. Converting this PR to a draft until this is implemented.