@jpountz 2024-08-02
Superseded by #13631.